reorganize containers and components
