+++ web/components/frontEndComponents/button/SecondaryButton/SecondaryButton.sass
@@ -1,4 +1,4 @@
-button
+button.secondary
   background: linear-gradient(180deg, #FFFFFF 0%, #EDFFFC 100%)
   cursor: pointer
 
@@ -20,14 +20,14 @@ button
   /* Link text */
   color: #00BCBC
 
-button:hover
+button.secondary:hover
   /* Link text/Hover */
   border: 1px solid #00D5D5
   box-shadow: 0px 4px 6px rgba(0, 82, 174, 0.15)
 
   color: #00D5D5
 
-button:focus
+button.secondary:focus
   outline: none
 
   /* Link text/Hover */
