+++ web/components/frontEndComponents/index.js
@@ -7,7 +7,7 @@ const SecondaryButtonCallback = () => console.log('Secondary Button clicked!');
 const FrontEndComponents = () => (
   <div>
     <p>Front end components!</p>
-    <SecondaryButton text='Secondary' onClick={SecondaryButtonCallback} />
+    <SecondaryButton text="Secondary" onClick={SecondaryButtonCallback} />
   </div>
 );
