--- web/containers/frontEndComponents/button/SecondaryButton/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import './SecondaryButton.sass';
-
-const SecondaryButton = ({ text, onClick }) => (
-  <button onClick={onClick}>
-    { text }
-  </button>
-);
-
-export default SecondaryButton;
