+++ web/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './containers/app/App';
+import App from './components/app/App';
 import './styles/index.sass';
 
 ReactDOM.render(<App />, document.getElementById('app'));
